7 урок, тесты. Добавил тесты в core_service

+++ cart_service/src/main/java/ru/geekbrains/spring/market/carts/model/CartItem.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -16,8 +18,8 @@ public class CartItem {
     private Long productId;
     private String productTitle;
     private int quantity;
-    private int pricePerProduct;
-    private int price;     // общая стоимость
+    private BigDecimal pricePerProduct;
+    private BigDecimal price;     // общая стоимость
 
 
 }
